Refactor land-grants folder structure
Organise files into controller,services, views folder structure rather than feature
